+++ workmanager_android/lib/workmanager_android.dart
@@ -66,7 +66,7 @@ class WorkmanagerAndroid extends WorkmanagerPlatform {
     Map<String, dynamic>? inputData,
     Duration? initialDelay,
     Constraints? constraints,
-    ExistingWorkPolicy? existingWorkPolicy,
+    ExistingPeriodicWorkPolicy? existingWorkPolicy,
     BackoffPolicy? backoffPolicy,
     Duration? backoffPolicyDelay,
     String? tag,