+++ workmanager_apple/lib/workmanager_apple.dart
@@ -66,7 +66,7 @@ class WorkmanagerApple extends WorkmanagerPlatform {
     Map<String, dynamic>? inputData,
     Duration? initialDelay,
     Constraints? constraints,
-    ExistingWorkPolicy? existingWorkPolicy,
+    ExistingPeriodicWorkPolicy? existingWorkPolicy,
     BackoffPolicy? backoffPolicy,
     Duration? backoffPolicyDelay,
     String? tag,